requirements.txt
discord.py~=2.4.0
flask
python-dotenv~=1.0.1
pillow~=11.0.0
watchdog~=6.0.0